refactor: FolderCell, FolderSelectorCell 통합
